backend_python/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
asgiref==3.8.1

astunparse==1.6.3
attrs==24.3.0
auth0-python==4.7.2
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.9.0
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.0
cattrs==24.1.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.8
cohere==5.13.11
coloredlogs==15.0.1

contourpy==1.3.1
crewai==0.95.0
crewai-tools==0.33.0
cryptography==43.0.3
cycler==0.12.1
dataclasses-json==0.6.7


Deprecated==1.2.15
deprecation==2.1.0
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docstring_parser==0.16
durationpy==0.9
embedchain==0.1.126
et_xmlfile==2.0.0
eventlet==0.39.0


fastapi==0.115.6
fastavro==1.10.0
ffmpeg-python==0.2.0
filelock==3.16.1
filetype==1.2.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.5.1
Flask-Sockets==0.2.1
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
gast==0.6.0
gevent==24.11.1
gevent-websocket==0.10.1
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.79.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-genai==0.2.2
google-generativeai==0.8.3
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gptcache==0.1.44
greenlet==3.1.1
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.69.0
grpcio-tools==1.70.0
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
h5py==3.12.1
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10

importlib_resources==6.5.2
instructor==1.7.2


ipywidgets==8.1.5
itsdangerous==2.2.0

Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
json_repair==0.35.0
jsonpatch==1.33
jsonpickle==4.0.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1


jupyterlab_widgets==3.0.13
keras==3.8.0
kiwisolver==1.4.8
kubernetes==31.0.0
lancedb==0.18.0
langchain==0.3.17
langchain-cohere==0.3.5
langchain-community==0.3.16
langchain-core==0.3.33
langchain-experimental==0.3.4
langchain-google-genai==2.0.9
langchain-openai==0.2.14
langchain-text-splitters==0.3.5
langsmith==0.1.147
libclang==18.1.1
litellm==1.59.0
llama-cloud==0.1.11
llama-index==0.12.14
llama-index-agent-openai==0.4.3
llama-index-cli==0.4.0
llama-index-core==0.12.14
llama-index-embeddings-openai==0.3.1
llama-index-embeddings-vertex==0.3.1
llama-index-indices-managed-llama-cloud==0.6.4
llama-index-llms-openai==0.3.14
llama-index-llms-vertex==0.4.2
llama-index-multi-modal-llms-openai==0.4.2
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.4
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-vertexaivectorsearch==0.2.0
llama-parse==0.5.20
lxml==5.3.0
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.0
matplotlib==3.10.0

mdurl==0.1.2
mem0ai==0.1.48
ml-dtypes==0.4.1
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
namex==0.0.8

networkx==3.4.2
nltk==3.9.1
nodeenv==1.9.1
numpy
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.59.7
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
opt_einsum==3.4.0
optree==0.14.0
orjson==3.10.15
overrides==7.7.0

pandas==2.2.3

pdfminer.six==20231228
pdfplumber==0.11.5


pillow==11.1.0

portalocker==2.10.1
posthog==3.8.3

propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3



pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyav==14.0.1
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydub==0.25.1

PyJWT==2.10.1
pylance==0.22.0
PyMySQL==1.1.1
pyparsing==3.2.1
pypdf==5.2.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyright==1.1.393
pysbd==0.3.4

python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.20
python-socketio==5.12.1
pytube==15.0.0
pytz==2024.2
pyvis==0.3.2
PyYAML==6.0.2

qdrant-client==1.13.2
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
requests-cache==1.2.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
schema==0.7.7
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
setuptools==75.1.0
shapely==2.0.6
shellingham==1.5.4
simple-websocket==1.1.0

sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37

starlette==0.41.3
striprtf==0.0.26
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.3
tomli==2.2.1
tomli_w==1.2.0

tqdm==4.67.1

typer==0.15.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0

tzdata==2024.2
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==2.3.0
uv==0.5.21
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4

websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
wheel==0.44.0
widgetsnbextension==4.0.13
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
yt-dlp==2025.1.26

zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
